making installable package
