+++ APT/APTB_extension.py
@@ -10,17 +10,11 @@ import pint.random_models
 from pint.phase import Phase
 from pint.fitter import WLSFitter
 from copy import deepcopy
-from astropy import log
 import astropy.units as u
 import numpy as np
 import matplotlib.pyplot as plt
-import os
-import csv
-import operator
-import time
 from pathlib import Path
-import socket
-import APTB
+from APT import APTB
 import treelib
 import matplotlib as mpl
